+++ src/Firebase/Http/HttpClientOptions.php
@@ -10,10 +10,8 @@ final class HttpClientOptions
 {
     /**
      * The amount of seconds to wait while connecting to a server.
-     *
-     * Defaults to indefinitely.
      */
-    private ?float $connectTimeout = null;
+    private ?float $connectTimeout = 15;
 
     /**
      * The amount of seconds to wait while reading a streamed body.
@@ -24,10 +22,8 @@ final class HttpClientOptions
 
     /**
      * The amount of seconds to wait for a full request (connect + transfer + read) to complete.
-     *
-     * Defaults to indefinitely.
      */
-    private ?float $timeout = null;
+    private ?float $timeout = 30;
 
     /**
      * The proxy that all requests should be passed through.